improved file rename

+++ code/image_library.rb
@@ -33,11 +33,13 @@ class ImageLibrary
   def change_name(new_name)
     full_old_name = @files[@index]
     full_new_name = @dir+new_name
+    return false if File.exist?(full_old_name)
     if full_old_name != full_new_name
       File.rename(full_old_name, full_new_name)
       @files[@index] = full_new_name
       puts "'#{full_old_name}' renamed '#{new_name}'"
     end
+    true
   end
 
   private
@@ -64,7 +66,5 @@ class ImageLibrary
     @index == 0 ? @index = @files.count - 1 : @index -= 1
     @full_file_name = @files[@index]
   end
-
-
 
 end
